# faster Toom-Cook 3 algorithms