@minad 2019-05-19
If what you are saying about the cutoff points, this is good :+1:

How is test coverage? Maybe add a tiny test specifically calling s_mul_karatsuba and s_mul_toom and comparing for very large numbers?