@czurnieden 2019-05-19
> If what you are saying about the cutoff points, this is good

I have updated my home-machine, if you remember and since then, some things work&hellip;a bit differently, to say the least.
The multiplication algorithms work as expected but the cut-off points of the squaring algorithms went through the roof (Tom's original ones and mine).
First observation: no swapping; the swap partition stays empty and I mean *empty*, zero bytes.
So Linux Mint 19.1 is optimized for SSDs. Will need to check if Ubuntu Bionic has the same defaults and might need to act accordingly.

And no, I don't know yet why it is squaring only but I have some suspicions.

Can you do me the favour and, if you have access to Bionic (virtual won't work here), run `tune`?

There is also the patch for this CPU branching problem (Meltdown, spectre, or whatever they are called) but Mint's repository offers no kernel without the patches, have to build one myself and that will take some time. But I doubt it will have a large effect here.

> Maybe add a tiny test specifically calling s_mul_karatsuba and s_mul_toom and comparing for very large numbers?

`tune.c` has testing abilities build in (just a `mp_cmp` against already generated values). I would like to call it for that purpose instead of mirroring the relevant code in `test.c` but I need to take it apart for it.

Was already working at it ( it's all in `main`, it's quite some spaghetti code and in urgent need of refactoring ;-) ) when I stumbled over the problem described above.
Quite a WTF moment, I can tell you!