@czurnieden 2019-05-20
Got to something acceptable by reducing the amount of variables to one, but it is still not known what the precise cause is.

New version:
```text
> summary(data)
       km               ks             tc3m            tc3s      
 Min.   : 85.00   Min.   :111.0   Min.   :167.0   Min.   :162.0  
 1st Qu.: 92.00   1st Qu.:126.0   1st Qu.:175.0   1st Qu.:200.8  
 Median : 93.00   Median :128.0   Median :178.0   Median :213.0  
 Mean   : 93.83   Mean   :125.8   Mean   :178.7   Mean   :213.0  
 3rd Qu.: 97.00   3rd Qu.:128.0   3rd Qu.:182.0   3rd Qu.:220.0  
 Max.   :104.00   Max.   :128.0   Max.   :191.0   Max.   :253.0  
> apply(data, 2, sd)
       km        ks      tc3m      tc3s 
 3.544964  4.493261  5.179710 14.389895
```