@czurnieden 2019-05-21
> This is also still work in progress?

I'm waiting for the approval&merging of #280 to rebase on that.
What's the problem with #280?
Oh.
\*blushes\*

>  I saw you mentioned gprof - I recommend perf these days

Am just used to it. Don't need more than a short glimpse at the output to see what the problem/bottleneck is. And recompiling? Don't use it very often, once or twice a year max, so that doesn't matter.