@minad 2019-05-21
> With `git merge` it is `-Xours` but with `git rebase` it is `-Xtheirs`. I hope I've chosen the right one ;-)
> 
> Oh, that dreaded `callgraph.txt` again *grrr*