@czurnieden 2019-05-21
> I thought before about removing [`callgraph.txt`]

It is a "nice to have" but I'm pretty sure any of the better IDEs around are able to it and do it much better. Should be generated on demand only.

I'll take a look at it later.