@minad 2019-05-21
> It is a "nice to have" but I'm pretty sure any of the better IDEs around are able to it and do it much better. Should be generated on demand only.

I would keep generating it and just gitignore it and remove it from git. But we also have tommath_class.h which is also bad, but not as bad.