@czurnieden 2019-05-21
> I would keep generating it and just gitignore it and remove it from git

Was quickly done, see #283 

> But we also have tommath_class.h which is also bad, but not as bad.

Yeah, true.
But that needs a bit more work to get rid off.