@sjaeckel 2019-05-21
@minad fine for you to merge even with the multi-clears instead of a single one?