@czurnieden 2019-05-21
> even with the multi-clears instead of a single one?

There is no `mp_init_size_multi`?