@sjaeckel 2019-05-21
> There is no `mp_init_size_multi`?

probably `mp_init_multi()` and `mp_grow()`?