@czurnieden 2019-05-21
> probably mp_init_multi() and mp_grow()?

That was 5 seconds later, max! ;-)
OK \*sigh\*