@minad 2019-05-21
> @minad fine for you to merge even with the multi-clears instead of a single one?

I really don't like it since this is very error prone if something is changed. What is the problem in changing it?

I don't want to slow things down here, but if you want to merge things - there are quite a few PRs which are ready, e.g. #279, #276, #271, #275.