@minad 2019-05-21
#266 seems also ready. Not that you misunderstand me - I fully approve this PR, but no hurry - until two hours ago  @czurnieden still seemed to be doing performance optimization and profiling.