@czurnieden 2019-05-21
> until two hours ago

That was two hours ago and just rebasing. Performance optimization, or better: make it work at all under the new memory managment, has been done yesterday.

> What is the problem in changing it?

It's work! ;-)
(I'm already at it, may need a minute or two)