@czurnieden 2019-05-21
> What is the problem in changing it?

`mp_grow` is about 10% slower

```text
With mp_grow
#define MP_DEFAULT_KARATSUBA_MUL_CUTOFF 95
#define MP_DEFAULT_KARATSUBA_SQR_CUTOFF 127
#define MP_DEFAULT_TOOM_MUL_CUTOFF      180
#define MP_DEFAULT_TOOM_SQR_CUTOFF      238

With mp_init_size
#define MP_DEFAULT_KARATSUBA_MUL_CUTOFF 96
#define MP_DEFAULT_KARATSUBA_SQR_CUTOFF 125
#define MP_DEFAULT_TOOM_MUL_CUTOFF      158
#define MP_DEFAULT_TOOM_SQR_CUTOFF      216
```

Still change to init_mult/grow?