@sjaeckel 2019-05-21
> Still change to init_mult/grow?

IMO no