@czurnieden 2019-05-21
> >   Still change to init_mult/grow?

> IMO no

I know that realloc is a bit slower than a fresh allocation, but it is quite a lot, indeed!