@minad 2019-05-21
If allocations are really an issue and mp_init_size_multi would solve that, why not consider that?