@sjaeckel 2019-05-21
> Just a technical question - I wonder about that. What is causing the slowdown? More allocations? If you access the malloc cache I wouldn't expect such a big difference. I had expect that all the time is eaten up by the numerics code.

> I know that realloc is a bit slower than a fresh allocation, but it is quite a lot, indeed!

okay, if both of you are wondering about this I'll merge later