@czurnieden 2019-05-22
The Toom-Cook algorithms do not need a lot of small allocation, the numbers are quite large, so your 5% is probably all you can get without a specialized memory management.
GlibC's one got quite good over the last 10 years, I would have guessed that you would gain much more with a speed optimized malloc implementation.

Tomfastmath could make much better use of an inbuild memory management, but there is already a discussion about that if I remember it correctly.

But I did as you told me and run some tracing programs. The most useful for C&P'ing the results here is probably `ltrace`, the others (strace and valgrind with the tools callgrind and cachegrind) have either too little or way too much information. There are more programs, I know, but I'm still in the "bracketing" stage, reducing the problem set, so ltrace is sufficient.

The first four lines, the ones with the `#define`s, are the output of `make tune`, ltrace gets a fixed-rounds `tune` to look at.

Single rounds with the version of tc3mul and tc3qsr I started this branch with.
Optimization -O3
```text
#define MP_DEFAULT_KARATSUBA_MUL_CUTOFF 97
#define MP_DEFAULT_KARATSUBA_SQR_CUTOFF 126
#define MP_DEFAULT_TOOM_MUL_CUTOFF      189
#define MP_DEFAULT_TOOM_SQR_CUTOFF      892 <

$ ltrace -S -c ./tune -L 3 -r 10 -M 100 -m 11
% time     seconds  usecs/call     calls      function
------ ----------- ----------- --------- --------------------
 26.92    0.266543          63      4228 memset
 24.36    0.241186          63      3792 free
 24.04    0.238011          62      3792 calloc
 16.93    0.167600          63      2620 realloc
  6.66    0.065979          62      1050 memcpy
  0.92    0.009141          63       144 clock_gettime
  0.05    0.000482         120         4 __printf_chk
  0.03    0.000262          65         4 __errno_location
  0.03    0.000262          65         4 strtol
  0.01    0.000112          18         6 SYS_brk
  0.01    0.000103          20         5 SYS_mmap
  0.01    0.000089          22         4 SYS_write
  0.01    0.000081          20         4 SYS_mprotect
  0.01    0.000061          20         3 SYS_access
  0.00    0.000048          16         3 SYS_fstat
  0.00    0.000047          23         2 SYS_openat
  0.00    0.000029          14         2 SYS_close
  0.00    0.000026          26         1 SYS_munmap
  0.00    0.000019          19         1 SYS_read
  0.00    0.000014          14         1 SYS_arch_prctl
------ ----------- ----------- --------- --------------------
100.00    0.990095                 15670 total
```
As you can see: TC3-sqr is quite an outlier

Single rounds with the version of tc3mul and tc3qsr I started this branch with
Optimization -O2
(The timings are not really relevant here, the number and kind of calls is)

```text
#define MP_DEFAULT_KARATSUBA_MUL_CUTOFF 160 <
#define MP_DEFAULT_KARATSUBA_SQR_CUTOFF 128
#define MP_DEFAULT_TOOM_MUL_CUTOFF      228
#define MP_DEFAULT_TOOM_SQR_CUTOFF      255

$ ltrace -S -c ./tune -L 3 -r 10 -M 100 -m 11
% time     seconds  usecs/call     calls      function
------ ----------- ----------- --------- --------------------
 36.76    0.232124          61      3792 calloc
 36.46    0.230235          60      3792 free
 25.12    0.158633          60      2620 realloc
  1.41    0.008935          62       144 clock_gettime
  0.07    0.000463         115         4 __printf_chk
  0.04    0.000271          67         4 __errno_location
  0.04    0.000265          66         4 strtol
  0.02    0.000116          19         6 SYS_brk
  0.01    0.000092          18         5 SYS_mmap
  0.01    0.000086          21         4 SYS_write
  0.01    0.000074          18         4 SYS_mprotect
  0.01    0.000063          21         3 SYS_access
  0.01    0.000048          24         2 SYS_openat
  0.01    0.000045          15         3 SYS_fstat
  0.00    0.000027          13         2 SYS_close
  0.00    0.000024          24         1 SYS_munmap
  0.00    0.000018          18         1 SYS_read
  0.00    0.000014          14         1 SYS_arch_prctl
------ ----------- ----------- --------- --------------------
100.00    0.631533                 10392 total
```
Here Kara-mul is the outlier. With the exception of Kara-mul it is also the result I got with my old system. (got me a Glibc 2.23 compiled \*pooh\*, no difference. Will try it with an old kernel later)

With the current versions and `-O3`

Single rounds mp_init_size

```text
$ ltrace -S -c  ./tune -L 3 -r 10 -M 100 -m 11
% time     seconds  usecs/call     calls      function
------ ----------- ----------- --------- --------------------
 30.62    0.313655          67      4615 memset
 21.00    0.215102          68      3145 realloc
 20.16    0.206486          67      3072 free
 20.14    0.206261          67      3072 calloc
  6.93    0.070965          67      1050 memcpy
  0.96    0.009820          68       144 clock_gettime
  0.05    0.000508         127         4 __printf_chk
  0.04    0.000402         100         4 __errno_location
  0.04    0.000387          96         4 strtol
  0.01    0.000121          20         6 SYS_brk
  0.01    0.000108          21         5 SYS_mmap
  0.01    0.000089          22         4 SYS_mprotect
  0.01    0.000086          21         4 SYS_write
  0.01    0.000067          22         3 SYS_access
  0.01    0.000055          18         3 SYS_fstat
  0.00    0.000046          23         2 SYS_openat
  0.00    0.000035          17         2 SYS_close
  0.00    0.000029          29         1 SYS_munmap
  0.00    0.000019          19         1 SYS_read
  0.00    0.000019          19         1 SYS_arch_prctl
------ ----------- ----------- --------- --------------------
100.00    1.024260                 15142 total
```

And with mp_grow
```text

$ ltrace -S -c ./tune -L 3 -r 10 -M 100 -m 11
% time     seconds  usecs/call     calls      function
------ ----------- ----------- --------- --------------------
 28.81    0.264157          66      3945 memset
 21.88    0.200644          65      3072 free
 21.85    0.200405          65      3072 calloc
 18.54    0.170003          68      2475 realloc
  7.66    0.070236          66      1050 memcpy
  1.04    0.009559          66       144 clock_gettime
  0.05    0.000501         125         4 __printf_chk
  0.05    0.000431         107         4 __errno_location
  0.04    0.000388          97         4 strtol
  0.01    0.000124          20         6 SYS_brk
  0.01    0.000120          24         5 SYS_mmap
  0.01    0.000096          24         4 SYS_mprotect
  0.01    0.000079          19         4 SYS_write
  0.01    0.000074          24         3 SYS_access
  0.01    0.000054          18         3 SYS_fstat
  0.01    0.000053          26         2 SYS_openat
  0.00    0.000036          18         2 SYS_close
  0.00    0.000032          32         1 SYS_munmap
  0.00    0.000023          23         1 SYS_read
  0.00    0.000018          18         1 SYS_arch_prctl
------ ----------- ----------- --------- --------------------
100.00    0.917033                 13802 total
```

Current version with LTM -O2, tune -O3

```text
#define MP_DEFAULT_KARATSUBA_MUL_CUTOFF 162 <
#define MP_DEFAULT_KARATSUBA_SQR_CUTOFF 128
#define MP_DEFAULT_TOOM_MUL_CUTOFF      191 <
#define MP_DEFAULT_TOOM_SQR_CUTOFF      202

% time     seconds  usecs/call     calls      function
------ ----------- ----------- --------- --------------------
 33.86    0.209098          66      3145 realloc
 32.20    0.198796          64      3072 calloc
 32.18    0.198707          64      3072 free
  1.49    0.009174          63       144 clock_gettime
  0.08    0.000482         120         4 __printf_chk
  0.06    0.000355          88         4 __errno_location
  0.05    0.000295          73         4 strtol
  0.02    0.000102          17         6 SYS_brk
  0.01    0.000087          21         4 SYS_write
  0.01    0.000080          16         5 SYS_mmap
  0.01    0.000067          16         4 SYS_mprotect
  0.01    0.000049          16         3 SYS_access
  0.01    0.000043          14         3 SYS_fstat
  0.01    0.000033          16         2 SYS_openat
  0.01    0.000033          16         2 SYS_close
  0.00    0.000023          23         1 SYS_munmap
  0.00    0.000014          14         1 SYS_read
  0.00    0.000013          13         1 SYS_arch_prctl
------ ----------- ----------- --------- --------------------
100.00    0.617451                  9477 total
```
Same with  -O2 (both)

```text
#define MP_DEFAULT_KARATSUBA_MUL_CUTOFF 156 <
#define MP_DEFAULT_KARATSUBA_SQR_CUTOFF 128
#define MP_DEFAULT_TOOM_MUL_CUTOFF      193 <
#define MP_DEFAULT_TOOM_SQR_CUTOFF      195

% time     seconds  usecs/call     calls      function
------ ----------- ----------- --------- --------------------
 33.26    0.207328          65      3145 realloc
 32.57    0.203006          66      3072 calloc
 32.40    0.201967          65      3072 free
  1.48    0.009204          63       144 clock_gettime
  0.08    0.000483         120         4 __printf_chk
  0.06    0.000377          94         4 __errno_location
  0.05    0.000306          76         4 strtol
  0.02    0.000109          18         6 SYS_brk
  0.02    0.000101          20         5 SYS_mmap
  0.01    0.000083          20         4 SYS_mprotect
  0.01    0.000077          19         4 SYS_write
  0.01    0.000062          20         3 SYS_access
  0.01    0.000049          16         3 SYS_fstat
  0.01    0.000044          22         2 SYS_openat
  0.01    0.000032          16         2 SYS_close
  0.00    0.000028          28         1 SYS_munmap
  0.00    0.000017          17         1 SYS_read
  0.00    0.000017          17         1 SYS_arch_prctl
------ ----------- ----------- --------- --------------------
100.00    0.623290                  9477 total
```
No difference between [LTM -O2, tune -O3] and both -O2, all well inside the error-bars. Would have made me wonder if not but now I have it that `tune` itself does not influence the result significantly.

The most significant outlier is TC3sqr within the very first measurement, so I'll start from there.